test(vetting): align preserve-hu tests with the oracle-vetting taxonomy
Reconcile PR 369 (--preserve-hu) with the vetting framework this branch introduces:

- check_coverage.py: add `pydicom` to ALLOWED_ORACLES so the validator matches the
  SPEC 4 catalog (prevents a latent CI failure once pydicom appears in the registry).
- Merge test_hounsfield_loader.h into test_hu_mechanics.h: the loader I/O checks are
  plumbing (SPEC 2 mechanics), and `_loader` is not a taxonomy suffix (SPEC 6.1). All
  HU mechanics now live in one file; drop the include and fix the fixture README ref.
- --preserve-hu adds no new feature (it swaps min-max rescaling for a slope-1 offset),
  so no new registry rows: record the pydicom oracle as corroboration on the existing
  MIN/MAX/MEAN/INTEGRATED_INTENSITY firstorder rows under a new firstorder.preserve_hu
  config recipe, per the established corroboration-in-notes convention.
